Validasi data Agama, Akun dan Bill

+++ app/Http/Controllers/BillController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use App\Helper\AlertHelper;
 use App\Models\Bills;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Crypt;
 use Illuminate\Support\Facades\DB;
 
@@ -21,16 +22,21 @@ class BillController extends Controller
      */
     public function index()
     {
-        $bills = Bills::orderBy('id', 'ASC')->get();
-        $data = [
-            'title' => $this->title,
-            'menu' => $this->menu,
-            'submenu' => $this->submenu,
-            'label' => 'data Tagihan',
-            'bills' => $bills
-        ];
+        $session_menu = explode(',', Auth::user()->akses_submenu);
+        if (in_array('27', $session_menu)) {
 
-        return view('bills.index')->with($data);
+            $bills = Bills::orderBy('id', 'ASC')->get();
+            $data = [
+                'title' => $this->title,
+                'menu' => $this->menu,
+                'submenu' => $this->submenu,
+                'label' => 'data Tagihan',
+                'bills' => $bills
+            ];
+            return view('bills.index')->with($data);
+        } else {
+            return view('not_found');
+        }
     }
 
     /**
@@ -40,13 +46,18 @@ class BillController extends Controller
      */
     public function create()
     {
-        $data = [
-            'title' => $this->title,
-            'menu' => $this->menu,
-            'submenu' => $this->submenu,
-            'label' => 'tambah Tagihan',
-        ];
-        return view('bills.create')->with($data);
+        $session_menu = explode(',', Auth::user()->akses_submenu);
+        if (in_array('28', $session_menu)) {
+            $data = [
+                'title' => $this->title,
+                'menu' => $this->menu,
+                'submenu' => $this->submenu,
+                'label' => 'tambah Tagihan',
+            ];
+            return view('bills.create')->with($data);
+        } else {
+            return view('not_found');
+        }
     }
 
     /**
@@ -61,21 +72,27 @@ class BillController extends Controller
             'bills' => 'required|unique:bills,bills,NULL,id,deleted_at,NULL',
             'notes' => 'required'
         ]);
-        DB::beginTransaction();
-        try {
-            Bills::create([
-                'bills' => $validated['bills'],
-                'notes' => $validated['notes'],
-                'looping' => isset($request->looping) ? 1 : 0
-            ]);
-            DB::commit();
-            AlertHelper::addAlert(true);
-            return redirect('bills');
-        } catch (\Throwable $err) {
-            DB::rollBack();
-            throw $err;
-            AlertHelper::addAlert(false);
-            return back();
+        $session_menu = explode(',', Auth::user()->akses_submenu);
+        if (in_array('28', $session_menu)) {
+
+            DB::beginTransaction();
+            try {
+                Bills::create([
+                    'bills' => $validated['bills'],
+                    'notes' => $validated['notes'],
+                    'looping' => isset($request->looping) ? 1 : 0
+                ]);
+                DB::commit();
+                AlertHelper::addAlert(true);
+                return redirect('bills');
+            } catch (\Throwable $err) {
+                DB::rollBack();
+                throw $err;
+                AlertHelper::addAlert(false);
+                return back();
+            }
+        } else {
+            return view('not_found');
         }
     }
 
@@ -98,15 +115,20 @@ class BillController extends Controller
      */
     public function edit($id)
     {
-        $bills = Bills::findOrFail(Crypt::decryptString($id));
-        $data = [
-            'title' => $this->title,
-            'menu' => $this->menu,
-            'submenu' => $this->submenu,
-            'label' => 'Ubah Tagihan',
-            'bills' => $bills
-        ];
-        return view('bills.edit')->with($data);
+        $session_menu = explode(',', Auth::user()->akses_submenu);
+        if (in_array('29', $session_menu)) {
+            $bills = Bills::findOrFail(Crypt::decryptString($id));
+            $data = [
+                'title' => $this->title,
+                'menu' => $this->menu,
+                'submenu' => $this->submenu,
+                'label' => 'Ubah Tagihan',
+                'bills' => $bills
+            ];
+            return view('bills.edit')->with($data);
+        } else {
+            return view('not_found');
+        }
     }
 
     /**
@@ -118,26 +140,31 @@ class BillController extends Controller
      */
     public function update(Request $request, $id)
     {
-        $decrypted_id = Crypt::decryptString($id);
-        $validated = $request->validate([
-            'bills' => "required|unique:bills,bills,$decrypted_id,id,deleted_at,NULL",
-            'notes' => 'required'
-        ]);
+        $session_menu = explode(',', Auth::user()->akses_submenu);
+        if (in_array('29', $session_menu)) {
+            $decrypted_id = Crypt::decryptString($id);
+            $validated = $request->validate([
+                'bills' => "required|unique:bills,bills,$decrypted_id,id,deleted_at,NULL",
+                'notes' => 'required'
+            ]);
 
-        DB::beginTransaction();
-        try {
-            $special_need = Bills::findOrFail($decrypted_id);
-            $special_need->bills = $validated['bills'];
-            $special_need->notes = $validated['notes'];
-            $special_need->looping = isset($request->looping) ? 1 : 0;
-            $special_need->save();
-            DB::commit();
-            AlertHelper::updateAlert(true);
-            return redirect('bills');
-        } catch (\Throwable $err) {
-            DB::rollBack();
-            AlertHelper::updateAlert(false);
-            return back();
+            DB::beginTransaction();
+            try {
+                $special_need = Bills::findOrFail($decrypted_id);
+                $special_need->bills = $validated['bills'];
+                $special_need->notes = $validated['notes'];
+                $special_need->looping = isset($request->looping) ? 1 : 0;
+                $special_need->save();
+                DB::commit();
+                AlertHelper::updateAlert(true);
+                return redirect('bills');
+            } catch (\Throwable $err) {
+                DB::rollBack();
+                AlertHelper::updateAlert(false);
+                return back();
+            }
+        } else {
+            return view('not_found');
         }
     }
 
@@ -149,17 +176,22 @@ class BillController extends Controller
      */
     public function destroy($id)
     {
-        DB::beginTransaction();
-        try {
-            $bills = Bills::findOrFail(Crypt::decryptString($id));
-            $bills->delete();
-            DB::commit();
-            AlertHelper::deleteAlert(true);
-            return back();
-        } catch (\Throwable $err) {
-            DB::rollBack();
-            AlertHelper::deleteAlert(false);
-            return back();
+        $session_menu = explode(',', Auth::user()->akses_submenu);
+        if (in_array('29', $session_menu)) {
+            DB::beginTransaction();
+            try {
+                $bills = Bills::findOrFail(Crypt::decryptString($id));
+                $bills->delete();
+                DB::commit();
+                AlertHelper::deleteAlert(true);
+                return back();
+            } catch (\Throwable $err) {
+                DB::rollBack();
+                AlertHelper::deleteAlert(false);
+                return back();
+            }
+        } else {
+            return view('not_found');
         }
     }
 }
